Make image load from filepath
This commit makes the image inserted into the PDF
load from a filepath instead of a PIL buffer so
optimizations baked into the image are handled
properly by reportlab.

In short, it seems that the version of RL in use
would not handle PIL buffers properly and end up
trashing optimizations, but by using the file,
the optimizations are retained.

Ref:
 * https://groups.google.com/g/reportlab-users/c/bwfCh0hIEWE/m/74tsbY2YAgAJ?pli=1
 * https://stackoverflow.com/a/75210218/2494976

Signed-off-by: Evan Gray <evanscottgray@gmail.com>